backend/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
django-cors-headers==4.9.0
nba_api==1.11.4
pandas==3.0.1
numpy==2.4.2
